pyproject.toml
[tool.black]
line-length = 88
target-version = ['py312']
exclude = '''
/(
    | migrations
)/
'''

[project]
name = "photos_app_backend"
version = "0.1.0"
description = "Backend for Photos App"
authors = [
    {name = "IT", email = "karkiaabishkar@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "Django>=5.0.2,<6.0.0",
    "djangorestframework>=3.14.0,<4.0.0",
    "dj-database-url>=2.1.0,<3.0.0",
    "psycopg2-binary>=2.9.9,<3.0.0",
    "asgiref>=3.7.2,<4.0.0",
    "uvicorn>=0.27.1,<0.28.0",
    "gunicorn>=21.2.0,<22.0.0",
    "importlib-metadata>=7.0.1,<8.0.0",
    "typing-extensions>=4.9.0,<5.0.0",
    "django-watchman>=1.3.0,<2.0.0",
    "drf-spectacular>=0.27.0,<0.28.0",
    "django-filter>=23.5,<24.0",
    "djangorestframework-simplejwt>=5.3.1,<6.0.0",
    "django-cors-headers>=4.3.1,<5.0.0",
    "django-extensions>=3.2.3,<4.0.0",
    "pillow>=11.1.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0.0",
    "coverage>=7.4.1,<8.0.0",
    "pytest-cov>=4.1.0,<5.0.0",
    "pytest-django>=4.8.0,<5.0.0",
    "ipdb>=0.13.13,<0.14.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel>=0.40.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]  # Explicitly define which packages to include

[tool.pytest.ini_options]
addopts = "--cov --cov-append --cov-report term-missing"
DJANGO_SETTINGS_MODULE = "app.settings"
python_files = "tests.py test_*.py *_tests.py"